License cleanups, trim down README.
In the simplest case, PLModel should probably assume that the user will
connect using postgres defaults, and so declaring connections should
not be necessary. The "for example" in the README should just show the
leanest version of code that will demonstrate the tidiness of the
feature set.

The license is clarified to refer to this particular project, and the
inflections test includes the copyright year, like Inflector.pm.