Add basic configuration to Database.
Configure the database connection through either Database::configure()
or passing arguments to the "use PLModel::Database" directive. Both
take a hash (not a hashref). Subsequent calls are additive; that is,
calling configure() once with a username and once with a password will
set both variables.

Configuration should be done before the initial connection is made.
The database will currently only reconnect if it disconnected.